+++ iop/network/smbman/src/ioman_add.h
@@ -21,8 +21,8 @@ typedef struct _iop_ext_device_ops {
 	int	(*close)(iop_file_t *);
 	int	(*read)(iop_file_t *, void *, int);
 	int	(*write)(iop_file_t *, void *, int);
-	int	(*lseek)(iop_file_t *, unsigned long, int);
-	int	(*ioctl)(iop_file_t *, unsigned long, void *);
+	int	(*lseek)(iop_file_t *, unsigned int, int);
+	int	(*ioctl)(iop_file_t *, unsigned int, void *);
 	int	(*remove)(iop_file_t *, const char *);
 	int	(*mkdir)(iop_file_t *, const char *);
 	int	(*rmdir)(iop_file_t *, const char *);
